Functional Prototype for both User and Restaurant Owners now.

+++ Megabyte/res_owner/views.py
@@ -1,24 +1,33 @@
 from django.shortcuts import render, redirect
-from .models import RestaurantOwner, Restaurant, Food, Category
+from .models import Restaurant, Food, Category
 from .forms import RestaurantForm, FoodForm, CategorizingForm, NewCategoryForm
+
+from accounts.models import CustomUser
+from user.models import Order
+
+from django.contrib.auth.decorators import login_required
 from django.http import Http404
 # Create your views here.
 
 
+@login_required
 def res_home_page(request):
     """
-    The home page for restaurant owners.
-    Show all restaurants the owners own in alphabetical order.
+    The home page for the restaurant owner
+    Show all the categories of a restaurant
     :param request: a HttpRequest object specific to Django
     :return the rendering of the html page 'res_owner/res_home_page.html'
     """
-    restaurants = Restaurant.objects.order_by('name')
-    # restaurants = Restaurant.objects.filter(owner=request.user).order_by('name')
-    # For when user is working
-    context = {"restaurants": restaurants}
+    this_user = CustomUser.objects.get(id=request.user.id)
+    # Make sure non-res-owners can't get access to the restaurant owner page
+    if not this_user.is_res_owner:
+        raise Http404
+    restaurants = this_user.restaurant_set.all()
+    context = {'restaurants': restaurants}
     return render(request, 'res_owner/res_home_page.html', context)
 
 
+@login_required
 def restaurant(request, restaurant_id: int):
     """
     The page for each restaurant.
@@ -28,13 +37,10 @@ def restaurant(request, restaurant_id: int):
     :return the rendering of the html page 'res_owner/restaurant.html'
     """
     this_restaurant = Restaurant.objects.get(id=restaurant_id)
-    foods = this_restaurant.food_set.all()
-
-    # Find all categories in all food of a restaurant.
-    res_category = set()
-    for food in foods:
-        for food_category in food.category_set.all():
-            res_category.add(food_category)
+    # Check to make sure different user cannot enter into other users' pages
+    if this_restaurant.restaurant_owner != request.user:
+        raise Http404
+    res_category = this_restaurant.category_set.all()
     sorted_category = sorted(list(res_category), key=lambda x: x.name)
     # Needs refactoring...also in the restaurant.html too
     context = {
@@ -46,6 +52,7 @@ def restaurant(request, restaurant_id: int):
     return render(request, 'res_owner/restaurant.html', context)
 
 
+@login_required
 def category(request, category_name: str, restaurant_id: int):
     """
     The page for each category.
@@ -57,12 +64,15 @@ def category(request, category_name: str, restaurant_id: int):
     """
     this_category = Category.objects.get(name=category_name)
     this_restaurant = Restaurant.objects.get(id=restaurant_id)
+    # Check to make sure different user cannot enter into other users' pages
+    if this_restaurant.restaurant_owner != request.user:
+        raise Http404
     all_foods = this_category.food.all()
     foods = []
     for food in all_foods:
         if food.restaurant.name == this_restaurant.name:
             foods.append(food)
-    restaurant_name = foods[0].restaurant.name
+    restaurant_name = this_restaurant.name
     sorted_food = sorted(foods, key=lambda a: a.name)
     context = {
         'foods': sorted_food,
@@ -73,6 +83,7 @@ def category(request, category_name: str, restaurant_id: int):
     return render(request, 'res_owner/category.html', context)
 
 
+@login_required
 def categorizing(request, category_name: str, restaurant_id: int):
     """
     The page for assigning a category to the foods
@@ -83,17 +94,22 @@ def categorizing(request, category_name: str, restaurant_id: int):
             or a redirect to the home page upon a successful POST request
     """
     this_category = Category.objects.get(name=category_name)
-
+    this_restaurant = Restaurant.objects.get(id=restaurant_id)
+    # Check to make sure different user cannot enter into other users' pages
+    if this_restaurant.restaurant_owner != request.user:
+        raise Http404
     if request.method != 'POST':
         # Initial request: Pre-fill form with the current entry
-        form = CategorizingForm(restaurant_id=restaurant_id, instance=this_category)
+        form = CategorizingForm(
+            restaurant_id=restaurant_id, instance=this_category)
     else:
         # POST request type confirmed, process data
-        form = CategorizingForm(data=request.POST, restaurant_id=restaurant_id, instance=this_category)
+        form = CategorizingForm(
+            data=request.POST, restaurant_id=restaurant_id, instance=this_category)
         if form.is_valid():
             form.save()
             if len(form.food_not_in_res) > 0:
-                existing_cat_name = form.data['name']
+                existing_cat_name = this_category.name
                 this_category = Category.objects.get(name=existing_cat_name)
                 for food in form.food_not_in_res:
                     this_category.food.add(food)
@@ -101,10 +117,12 @@ def categorizing(request, category_name: str, restaurant_id: int):
             return redirect('res_owner:res_home_page')
 
     # This sends the context to render the edit_restaurant.html
-    context = {'form': form, 'category_name': category_name, 'restaurant_id': restaurant_id}
+    context = {'form': form, 'category_name': category_name,
+               'restaurant_id': restaurant_id}
     return render(request, 'res_owner/categorizing.html', context)
 
 
+@login_required
 def new_category(request, restaurant_id: int):
     """
     The page for adding new categories + categorization of food items.
@@ -113,14 +131,17 @@ def new_category(request, restaurant_id: int):
     :return the rendering of the html page 'res_owner/new_category.html' upon a GET request
             or a redirect to the home page upon a successful POST request
     """
+    this_restaurant = Restaurant.objects.get(id=restaurant_id)
+    # Check to make sure different user cannot enter into other users' pages
+    if this_restaurant.restaurant_owner != request.user:
+        raise Http404
     if request.method != 'POST':
         # Empty Form
         form = NewCategoryForm(restaurant_id=restaurant_id)
     else:
         # POST request type confirmed, process data
         form = NewCategoryForm(data=request.POST, restaurant_id=restaurant_id)
-
-        # Check if there is an error associated with field name in the form.
+        # Check if there is an error associated with field name in the form. Cursed, I know.
         if form.errors.get('name') is not None:
             # When the duplicate key in Category table exists.
             if form.errors['name'].as_data():
@@ -131,22 +152,27 @@ def new_category(request, restaurant_id: int):
                     restaurant_id=restaurant_id,
                     instance=this_category
                 )
-
         if form.is_valid():
-            form.save()
+            # Save first then add later
+            new_category = form.save(commit=False)
+            new_category.save()
+            new_category.restaurant.add(this_restaurant)
             if len(form.food_not_in_res) > 0:
                 existing_cat_name = form.data['name']
                 this_category = Category.objects.get(name=existing_cat_name)
+                # Readd food from other restaurants to the Category.
+                # This is due to the MultipleChoiceField in NewCategory form that can cause foods from other restaurants
+                # to be uncategorized when only foods from this restaurant that can be categorized appears in that field
                 for food in form.food_not_in_res:
                     this_category.food.add(food)
                 this_category.save()
             return redirect('res_owner:res_home_page')
-
-    # This sends the context to render the edit_restaurant.html
+        # This sends the context to render the edit_restaurant.html
     context = {'form': form, 'restaurant_id': restaurant_id}
     return render(request, 'res_owner/new_category.html', context)
 
 
+@login_required
 def delete_category(request, category_name: str, restaurant_id: int):
     """
     The view function that handles the deletion of a category off a restaurant.
@@ -157,17 +183,24 @@ def delete_category(request, category_name: str, restaurant_id: int):
     :param category_name: the name of the category in the Category table
     :return a redirect to the home page upon a successful POST request
     """
+    this_restaurant = Restaurant.objects.get(id=restaurant_id)
+    # Check to make sure different user cannot enter into other users' pages
+    if this_restaurant.restaurant_owner != request.user:
+        raise Http404
+
     if request.method == 'POST':
-        this_restaurant = Restaurant.objects.get(id=restaurant_id)
         this_category = Category.objects.get(name=category_name)
         foods = this_restaurant.food_set.all()
         # Count the number of foods in this restaurant with this category
         foods_with_same_cat = 0
         for food in foods:
             for food_category in food.category_set.all():
+                # Remove every food associated with this category in this restaurant when deleting the category
                 if food_category.name == category_name:
                     this_category.food.remove(food)
                     foods_with_same_cat += 1
+        # Remove the restaurant associating with this category too
+        this_category.restaurant.remove(this_restaurant)
         # print(this_category.food.all())
         # print(foods_with_same_cat)
         # Basically, this means this category is used only by this restaurant. So delete it off the database also
@@ -176,6 +209,7 @@ def delete_category(request, category_name: str, restaurant_id: int):
     return redirect('res_owner:res_home_page')
 
 
+@login_required
 def cat_others(request, restaurant_id: int):
     """
        The page for category Others
@@ -185,6 +219,9 @@ def cat_others(request, restaurant_id: int):
        :return the rendering of the HTML page 'res_owner/category.html'
    """
     this_restaurant = Restaurant.objects.get(id=restaurant_id)
+    # Check to make sure different user cannot enter into other users' pages
+    if this_restaurant.restaurant_owner != request.user:
+        raise Http404
     all_foods = this_restaurant.food_set.all()
     no_category = []
     for food in all_foods:
@@ -201,21 +238,24 @@ def cat_others(request, restaurant_id: int):
     return render(request, 'res_owner/category.html', context)
 
 
+@login_required
 def res_settings(request):
     """
     The page for managing the restaurants
     :param request: a HttpRequest object specific to Django
     :return the rendering of the HTML page 'res_owner/res_settings.html'
     """
-    restaurants = Restaurant.objects.order_by('name')
-    # restaurants = Restaurant.objects.filter(owner=request.user).order_by('name')
-    # For when user is working
+    # Disallow non restaurant owners from accessing restaurant-owner specific pages
+    if request.user.is_res_owner is False:
+        raise Http404
+    restaurants = Restaurant.objects.filter(restaurant_owner=request.user).order_by('name')
     context = {
         'restaurants': restaurants
     }
     return render(request, 'res_owner/res_settings.html', context)
 
 
+@login_required
 def new_restaurant(request):
     """
     The page for adding in a new restaurant.
@@ -223,6 +263,10 @@ def new_restaurant(request):
     :return the rendering of the HTML page 'res_owner/new_restaurant.html' upon a GET request
             or a redirect to the home page upon a successful POST request
     """
+    # Disallow non restaurant owners from accessing restaurant-owner specific pages
+    if request.user.is_res_owner is False:
+        raise Http404
+
     if request.method != 'POST':
         # Blank form
         form = RestaurantForm()
@@ -232,13 +276,14 @@ def new_restaurant(request):
         # Might have to change it once custom form validation is implemented.
         if form.is_valid():
             new_restaurant = form.save(commit=False)
-            # new_restaurant.restaurant_owner = request.owner
+            new_restaurant.restaurant_owner = request.user
             new_restaurant.save()  # Save to the database
             return redirect('res_owner:res_home_page')
     context = {'form': form}
     return render(request, 'res_owner/new_restaurant.html', context)
 
 
+@login_required
 def edit_restaurant(request, restaurant_id: int):
     """
     The page for editing an existing restaurant entry.
@@ -248,10 +293,10 @@ def edit_restaurant(request, restaurant_id: int):
             or a redirect to the home page upon a successful POST request.
     """
     this_restaurant = Restaurant.objects.get(id=restaurant_id)
+    # Check to make sure different user cannot enter into other users' pages
+    if this_restaurant.restaurant_owner != request.user:
+        raise Http404
 
-    # Uncomment when user registration is completed
-    # if this_restaurant.restaurant_owner != request.user:
-    #     raise Http404
     if request.method != 'POST':
         # Initial request: Pre-fill form with the current entry
         form = RestaurantForm(instance=this_restaurant)
@@ -267,6 +312,7 @@ def edit_restaurant(request, restaurant_id: int):
     return render(request, 'res_owner/edit_restaurant.html', context)
 
 
+@login_required
 def delete_restaurant(request, restaurant_id: int):
     """
     The view function that handles the deletion of a restaurant
@@ -274,11 +320,16 @@ def delete_restaurant(request, restaurant_id: int):
     :param restaurant_id: the id of the restaurant in the Restaurant table to delete
     :return A redirect to the home page upon a successful POST request.
     """
+    this_restaurant = Restaurant.objects.get(id=restaurant_id)
+    # Check to make sure different user cannot enter into other users' pages
+    if this_restaurant.restaurant_owner != request.user:
+        raise Http404
     if request.method == 'POST':
         Restaurant.objects.filter(id=restaurant_id).delete()
     return redirect('res_owner:res_home_page')
 
 
+@login_required
 def new_food(request, restaurant_id: int):
     """
     The page for adding in a new food without categories.
@@ -287,6 +338,10 @@ def new_food(request, restaurant_id: int):
     :return The rendering of the HTML page 'res_owner/new_food.html' upon a GET request
             or a redirect to the home page upon a successful POST request.
     """
+    this_restaurant = Restaurant.objects.get(id=restaurant_id)
+    # Check to make sure different user cannot enter into other users' pages
+    if this_restaurant.restaurant_owner != request.user:
+        raise Http404
     if request.method != 'POST':
         # Blank form
         form = FoodForm()
@@ -294,7 +349,7 @@ def new_food(request, restaurant_id: int):
         this_restaurant = Restaurant.objects.get(id=restaurant_id)
         # POST request type confirmed; processing data
         form = FoodForm(data=request.POST)
-        if form.is_valid():
+        if form.is_valid() and not this_restaurant.food_set.filter(name=request.POST.get('name')):
             new_food = form.save(commit=False)
             new_food.restaurant = this_restaurant
             new_food.save()  # Save to the database
@@ -303,6 +358,7 @@ def new_food(request, restaurant_id: int):
     return render(request, 'res_owner/new_food.html', context)
 
 
+@login_required
 def edit_food(request, food_id: int):
     """
     The page for editing food info
@@ -312,9 +368,10 @@ def edit_food(request, food_id: int):
             or a redirect to the home page upon a successful POST request.
     """
     this_food = Food.objects.get(id=food_id)
-    # Uncomment when user registration is completed
-    # if this_food.restaurant.restaurant_owner != request.user:
-    #     raise Http404
+    # Check to make sure different user cannot enter into other users' pages
+    if this_food.restaurant.restaurant_owner != request.user:
+        raise Http404
+
     if request.method != 'POST':
         # Initial request: Pre-fill form with the current entry
         form = FoodForm(instance=this_food)
@@ -330,6 +387,7 @@ def edit_food(request, food_id: int):
     return render(request, 'res_owner/edit_food.html', context)
 
 
+@login_required
 def delete_food(request, food_id: int):
     """
     The page for deleting an existing restaurant entry.
@@ -337,6 +395,55 @@ def delete_food(request, food_id: int):
     :param food_id: the id of the food in the Food table to delete
     :return A redirect to the home page upon a successful POST request
     """
+    this_food = Food.objects.get(id=food_id)
+    this_restaurant = this_food.restaurant
+    # Check to make sure different user cannot enter into other users' pages
+    if this_food.restaurant.restaurant_owner != request.user:
+        raise Http404
+
     if request.method == 'POST':
         Food.objects.filter(id=food_id).delete()
-    return redirect('res_owner:res_home_page')
+    return redirect('res_owner:restaurant', restaurant_id=this_restaurant.id)
+
+
+# This might be a bad one....
+# This require order to have a foreign key to the restaurant.
+@login_required
+def order_management(request):
+    """
+    The page for viewing the list of orders for a restaurant.
+    :param request: a HttpRequest object specific to Django
+    :return the rendering of the HTML page orders.html
+    """
+    # Prevent normal user from accessing the restaurant owner pages
+    # Allowing res_owner to delete past order is QOL really....
+    if request.user.is_res_owner is False:
+        raise Http404
+    all_restaurants = request.user.restaurant_set.all()
+    orders = []
+    for a_restaurant in all_restaurants:
+        orders.append(list(a_restaurant.order_set.all()))
+    context = {'orders': orders}
+    return render(request, 'res_owner/orders.html', context)
+
+
+# WIP
+@login_required
+def change_order_status(request, order_id):
+    """
+    The page for changing the status of the order.
+    :param request: a HttpRequest object specific to Django
+    :param order_id: the id of the order in the Order table
+    """
+    this_order = Order.objects.get(id=order_id)
+    # Allow only the owner of this order's restaurant to see the order
+    if this_order.restaurants.restaurant_owner != request.user:
+        raise Http404
+    if request.method != 'POST':
+        pass
+    else:
+        pass
+    # The form should be similar to the radio button except you can only choose one.
+    available_statuses = ["Waiting", "Rejected", "Accepted", "Delivered"]
+    context = {'availabe_statuses': available_statuses}
+    return render(request, 'res_owner/change_order_status.html', context)
